+++ apps/roam/src/utils/advancedSearchFooterUtils.ts
@@ -1,6 +1,8 @@
 import getUids from "roamjs-components/dom/getUids";
 import getTextByBlockUid from "roamjs-components/queries/getTextByBlockUid";
 import updateBlock from "roamjs-components/writes/updateBlock";
+import getPageTitleByPageUid from "roamjs-components/queries/getPageTitleByPageUid";
+import openBlockInSidebar from "roamjs-components/writes/openBlockInSidebar";
 
 export type BlockSelection = {
   selectionStart: number;
@@ -16,6 +18,28 @@ export type InsertTarget = {
 };
 
 const DEFAULT_WINDOW_ID = "main-window";
+
+export const openSearchResultInMain = async (uid: string): Promise<void> => {
+  if (getPageTitleByPageUid(uid)) {
+    await window.roamAlphaAPI.ui.mainWindow.openPage({ page: { uid } });
+    return;
+  }
+  await window.roamAlphaAPI.ui.mainWindow.openBlock({ block: { uid } });
+};
+
+export const openSearchResultFromLinkEvent = async ({
+  shiftKey,
+  uid,
+}: {
+  shiftKey: boolean;
+  uid: string;
+}): Promise<void> => {
+  if (shiftKey) {
+    await openBlockInSidebar(uid);
+    return;
+  }
+  await openSearchResultInMain(uid);
+};
 
 export const getBlockSelection = (uid: string): BlockSelection => {
   const activeElement = document.activeElement;
